Cypher: support rewriting resolved functions
We use the `QuineFunctionInvocation` subtype of `FunctionInvocation` to
represent resolved functions. However, since we subclassed a `case
class`, it turns out that the default openCypher AST tree rewriting
machinery doesn't work properly: it rewrites the AST back to a regular
`FunctionInvocation`. The fix is simple: override the `dup` method to
manually specify the rewriting rules.

We haven't noticed so far because: most rewriting occurs _before_
function resolution and rewrites that modify a function argument are
rare. Still, they are possible, as visible from the regression test.

Fixes QU-433

GitOrigin-RevId: b67d662a90e2c1fda8c95c4669533ee56c044807