Static pass for identifying non-failing queries
This isn't used anywhere but it is something fun I think could come in
handy if refined further. In particular, we've discussed how errors in
Cypher SQ query outputs (or even filter/map in `MultipleValues` mode)
make for tricky debugging. This sort of static analysis pass could serve
as a check for warning the user that they need to specify some error
recovery strategy.

Closes out the final bit of QU-71

GitOrigin-RevId: 5a6f89bdbddea4c72ef0ebea8f3354c5728588f2