Synchronize `QuineApp` updates and `QuineEnterpriseApp` ingests on application configuration state
…rather than a var's (changing) value

This PR is best viewed as a "band-aid" fix -- this very aggressively synchronizes and Awaits to make correctness as clear as possible. This is ultimately stapling single-threading onto a native JVM object. Longer term, we want to rewrite this synchronization in terms of an orchestrating actor (or other natively single-threaded object). However, GraphServiceActor seems to have disappeared from the open source codebase, only living on in the enterprise code. Re-adding it is out of scope for the timeline of this PR.

GitOrigin-RevId: b0e3573d0eb11d61dea9c876e0f17946c013be6b