Add `blocking` to `synchronized` calls.
Per the scaladoc for this method: "Used to designate a piece of code
which potentially blocks, allowing the current BlockContext to adjust
the runtime's behavior. Properly marking blocking code may improve
performance or avoid deadlocks."

It's used in the impl of Await.ready / Await.result.

The threads that we're blocking with `synchronized` here are in
quine-cluster-akka.actor.default-dispatcher.

Without this change, running 10 concurrent create / delete ingests, I
would see 7 our of those 8 default dispatcher threads blocked in
"Monitor" status (light blue) on the `synchronized` in VisualVM, and
then the app was essentially wedged at that point - not logging anything
more, not responding to requests, etc.

With this change, more threads are created in that dispatcher as threads
are blocked. So I can send 100 concurrent create / delete ingests, and
I'll get 100 threads in that dispatcher, and everything works fine as
far as app correctness and responsiveness.

GitOrigin-RevId: bed89197b29b1430a8fa80d8fd4850fbc235cbe8