Fix possible hang in edge case of the REST API
Thanks to [an upstream bug][0] which is exposed when we quickly create
and delete and ingest stream (and possibly in other places - that's just
where the repro case came from), it is possibly for our REST API to
hang.

Of particular note: for clients that automatically retry idempotent HTTP
methods, this may manifest as the "DELETE ingest stream" API returning a
404 - the first call will hang + timeout but the second call will be
fine.

Fixes QU-651

[0]: https://github.com/akka/akka-stream-contrib/issues/119

GitOrigin-RevId: 88b89341afbedc1fc00ac2f2c408df66fe28b5ed