Don't include labels twice in literal operation
The `LiteralMessage.GetRawPropertiesCommand` includes labels twice: once
in the labels field and a second time as an entry in the properties map.
This changes that behaviour to omit the label from the properties.

This fixes #4 since `UserDefinedProcedure.getAsCypherNode` (which is
used by `recentNodes` and other procedures) now doesn't include the
label as a property.

Closes thatdot/quine#5

Co-authored-by: Alec Theriault <alec.theriault@gmail.com>
GitOrigin-RevId: c5a1832a37ce44075458d35cdd5ceb52ef974634