Retry appState.load in case of persistor failure
With this change, a persistence Failure during application load/re-hydration will cause the system to log a warning, wait 100 milliseconds, then retry hydration.

Note that we intentionally start the REST API potentially before the appState is loaded. This is so that the API can be used for status checks and liveness probes.

GitOrigin-RevId: 08898c6c361b3fff51919b752ef8ac8df6fdbed2