Recipes: accumulate substitution errors monadically
Instead of just throwing an exception when doing variable substitution
on recipes, this uses the `Validated` datatype from cats to aggregate
all of the various errors, primarily under an applicative context (so
that we collect as many errors as possible instead of aborting after the
first unbound variable).

The motivation for this is:

  - failed substitution prints out _all_ of the missing variables on the
    first run

  - it becomes possible to generate documentation about which variables
    a recipe expects (this is done inside `quine-docs`)

Also added a couple `quine-docs` improvements: the JAR name is now a
variable derived from `quine`'s version, YAML is syntax highlighted, GH
stats are hidden (since they don't render nicely).

GitOrigin-RevId: 171d3b0d854bdb8467ac33edd05d89c5f067fe78