Add gauge to track ingest valve state
There is now a gauge at "shared.valve.ingest" capturing the current
close/open state of the shared ingest valve. `n=0` means open, `n>0`
means the valve is closed in `n` places.

The ingest valve is used as a coarse-grain backpressuring mechanism.
However, if for some reason it ends up being permanently closed (eg.
because a standing query output failed or is stuck, or even just a bug),
this could be felt by the user as the system being stalled - ingest
would show up as "RUNNING", but still be stalled. This gauge makes it
easy to diagnose that situation.

GitOrigin-RevId: 071c6039ceb593fd6e16baf72bd961ab613f8b85