Cypher: keep hint in `Bytes` about ID provenance
In the Cypher interpreter, we represent IDs using byte arrays.
Unfortunately, this means that when IDs are exiting the interpreter,
there isn't enough information to differentiate raw byte arrays from
IDs. In order to rectify this situation, we can keep around one extra
boolean to track whether the byte array came from an ID and then use
that to decide how to format/emit the value when it comes out.

This renders obsolete the `StandingQuery.rawIdReturnColumns` hack - now
the `id(n)` columns will still be `Expr.Bytes`, but they'll have
`representsId = true` so will be turned properly back into a
`QuineValue.Id`.

GitOrigin-RevId: c77691cb31441e6fdbb53749cacd6eeedf880a97