Cassandra: add persistor tests and fix uncovered bugs
This adds a `PersistenceAgentSpec` test implementation for the Cassandra
persistor using the `embedded-cassandra` library.

This exposed two bugs

 * Comparison of `EventTime` is over the signed long when it should be
   unsigned. This is easy to fix (see `TypeCodec[EventTime]`), but it
   means `PersistenceAgent.CurrentVersion` must be bumped. It is
   unlikely anyone encountered this though - the overflow point for
   `EventTime` is 2039-09-07T15:47:35.552Z.

 * Journals are queried in descending event order when they should be
   ascending order. This would potentially mess up historical queries,
   but only if the journal event effects didn't commute. This also
   requires a change in schema.

Notes on `embedded-cassandra` and CI
===

Under the hood, this downloads the requested Cassandra version and
starts it in a fresh JVM. Besides not having a fast startup time, this
seems to be a pretty faithful test of Cassandra, which means that it
won't work on:

  - JDK that is too recent (at least not without the right JVM args,
    which I think I've added, but cannot test because...)
  - M1 since support for that is still forthcoming (due for 4.1)

As with RocksDB tests, the testsuite will be skipped if
`embedded-cassandra` fails to start. In CI, the testsuite will always
be run. Additionally, the ` ~/.embedded-cassandra` folder is cached (it
is about 60MB).

Co-authored-by: emanb29 <ethan@thatdot.com>
Co-authored-by: Leif Warner <abimelech@gmail.com>
GitOrigin-RevId: 1cec001cef013b32230f77c35fc5113a4bae8ebd