Re-add `hasEffect` check to processEvents with 1 event
The most visible effect of this bug was that MultipleValues that set a value they watched would infinitely cycle, even when the value was unchanged. This also would (slightly) decrease performance of DistinctId SQs, and cause some unnecessarily-long journals (as redundant events would have been recorded in the journal) and, if snapshot was set to `on-node-update`, unnecessary snapshots.

GitOrigin-RevId: c9fd2a1cc771c6f57247a4114a6e19feca504f10