Add fused "Return" cypher.Query AST node
This should have no user-facing changes -- all queries that compiled before should still compile, all that did not should still not. Some may compile to a different `cypher.Query` AST, but this should not be visible unless a user uses `EXPLAIN`.

Query results should also be unaffected -- all queries that returned results before should still return the same results. Some forms of query (those that use `ORDER BY` and `LIMIT` but not `SKIP` or `DISTINCT`) will be slightly more performant.

This commit also fixes `substitute` when used with partial Parameters by restructuring its arguments to better support "unspecified" values and use a more typical pattern for substitution

--

Background: A `RETURN` clause has 4 facets. In order, they are:
- `ORDER BY`
- `DISTINCT`
- `SKIP`
- `LIMIT`

A `RETURN` is one kind of projection available in Cypher ("projection" in the relational algebra sense). A projection may be aggregating or non-aggregating. Projections other than `RETURN` may also include a WHERE facet, but this PR focuses only on `RETURN`.

--

This commit adds a fused `cypher.Query.Return` Quine cypher AST node that tracks all 4 facets of a `RETURN` in a single `cypher.Query`. This new AST node is compiled from any `RETURN` clause in a query, but not from other kinds of projections (eg WITH clauses). Put another way: this creates an abstraction (`Query.Return`) over a set of related operators (`Sort`, `Distinct`, etc) that would otherwise have only a flat structure in queries.

When aggregations are not present (eg `count()`, `sum()`) a `RETURN` clause can be compiled to a single `Query.Return` with all information about how to handle that final projection. This allows the cypher interpreter to perform the RETURN with more optimizations, such as using less memory to perform an `ORDER BY ... LIMIT`

When aggregations _are_ present, due to how `ORDER BY` works (accessing scopes both "inside" and "outside" the return projection), the `RETURN` clause cannot include the `ORDER BY` specification. Instead, we compile a query that does the aggregation and ordering, then use that as an input to a `RETURN` managing the other 3 facets of a `RETURN` (i.e., `DISTINCT`/`SKIP`/`LIMIT`).

GitOrigin-RevId: 69224c7f8b3daa2f7f3fc2eb361d676b9a47e429