Remove unecessary Promise from GoToSleep behavior.
* Remove unecessary Promise from GoToSleep behavior.

Maybe a follow-up to https://github.com/thatdot/quine-plus/pull/1726,
but even before that, this Promise wasn't used - the message contained a
Future, which was the Future from the persistence actions.

```
p.completeWith {
  p.future.onComplete(doThings)
  f
}
```
is the same thing as:
```
f.onComplete(doThings)
```
afaik?
That's the main change done here.

* Made a helper method for the `send SleepSuccess` to the shard case, to
make it obvious that is being done two different places (after
persisting succeeds, or if no persisting is necessary).

* Removed Future.unit from the "no peristing necessary" case, as that
was unused (formerly that Future.unit was sent back to the shard in
GoingToSleep to let it know when persisting finished).

* Update description of sleep deadline

* Update comment

* scalafmt

* Move `retryPersistence` up a level and also use it in the snapshot case.

* Remove intermediate standing query serialzed collection.

We don't need the extra .view.map + .toVector; we can serialize + record
metrics in the same traversal over pendingStandingQueryWrites where
we're firing off the persistence Futures.

* Remove if case for empty collection.

Future.traverse is already a no-op for empty collections.
It'll just immediately return a Future(empty set) - nothing will be
dispatched to the ExecutionContext.

* Un-nest the standingQueryStatesSaved

* Remove zipWith - we don't care about the actual value of persitenceFuture

GitOrigin-RevId: 91a01f3e0fc2638a731c9f9361d4594ee21e41da