Inline and make explicit all ExecutionContexts
This contains no changes to where or how any tasks execute, except for the following:

- The NamedPipe source now uses our blocking dispatcher instead of the akka internal one.
- A few no-op tasks (`map (foo => foo)`) that were implicitly created by for comprehensions have been removed
- A few occurrences of mutually-compatible `flatMap`s that were implicitly created by for comprehensions have been replaced with explicit `zip` or `zipWith` for readability

Canonical references to dispatchers (and therefore ECs) have been encapsulated in a new `QuineDispatchers` utility class. I've also added notes to the quine-akka config files documenting the purpose and appropriate use-cases for each dispatcher.

The _very few_ (2 or 3) functions that really do need an EC as an argument have had that argument made explicit. Those occured for one of the following reasons:
1) different ECs were used by different invocations of the function
2) the function did not have a `graph` or `actorSystem` readily in scope, meaning it couldn't easily access a `QuineDispatchers` instance to retrieve the specific EC the function needs

A final note: This seems heavy-handed, and it is. Scala encourages its users to keep ECs implicit, to use only a single EC throughout an application, and to schedule everything against that EC. Quine requires a more nuanced approach to ECs. One motivating problem:

Even in the same scope, it's too easy to make a mistake with implicit ECs. We have this pattern all over our codebase:

`aFuture.map(doesSomeWork()).map(_ => ())`

The first `map` usually needs to be actually scheduled, while the `_ => ()` really shouldn't be independently scheduled.
There are 8 places in which we schedule literally that `_ => ()` task on _the shard dispatcher_. Making all ECs explicit makes it easy to catch mistakes like that.

GitOrigin-RevId: 1bb88ec88f22f7e6fa0811b81634ad2387bbd696