Fix check for unsigned long in msgpack
* Fix check for unsigned long in msgpack

* Factor out Instant serialization. This is in anticipation with that
code being shared by LocalDateTime serialization.
* Use msgpack's built-in timestamp extension type instead of writing our
own. One of the benefits originally given by us for use of msgpack is
that it's "standard", and useable via other tools. Using standard parts
of msgpack instead of doing our own thing facilitates that goal. Also,
it's less code for us to have. And it stores timestamps in 64 bits up to
the year 2514 (or 32 bits if the resolution is seconds), instead of our
96-bit long + int (though it still expands to the 96-bit representation
if needed for times outside that range). Left in deserialization for our
old format, so people can upgrade without noticing. Can be removed at
some future date. Though this will prevent people from rolling back, as
the old code can't read the new format.
https://github.com/msgpack/msgpack/blob/master/spec.md#timestamp-extension-type

* Remove our own to / from DateTime code.

Re-introduce this stuff for saving e.g. LocalDateTimes (stuffed in leif/datetime_types branch)

Update hashcodes in tests which depend on serialization format of
QuineValues.

GitOrigin-RevId: 5ff1265ee4e69f8317c37b0f3f66cd8d9899a6aa