Cypher: add text.urlencode and text.urldecode
The java notion of "URL encode" is really `x-www-form-urlencoded`, which deviates from javascript URI encoding in several ways, including that spaces are encoded as `+`. These functions allow the user to choose between browser-like (RFC3986) behavior, and java-like (x-www-form-urlencoded) behavior.

RFC3986's decoding is stricter than x-www-form-urlencoded, so we default to RFC3986 for encoding, but x-www-form-urlencoded for decoding, to maximize compatibility.

GitOrigin-RevId: aefdfa8e5928f299c1f1e51677e25f90a70b527a