Switch to snakeyaml-engine, supporting YAML 1.2
* Switch to snakeyaml-engine, supporting YAML 1.2

snakeyaml-engine is a sucessor of snakeyaml. The project is supposed to
be more modular - notably this jar doesn't include support for e.g.
mapping to Java Beans (which we don't use), and thus is smaller.

YAML 1.2 is a proper superset of JSON, and doesn't have the older YAML's
extra features like merging - and thus is easier to process.

Notably, support for hexadecimal-formatted numbers has been removed
(they're just interpreted as strings, now).

* Clarify snakeyaml-engine's JSON vs Core schema situation in comments.

GitOrigin-RevId: 19f5cbc21762ae60e60a7c70df2fe5ec01799dc2