QU-896: Catch InvalidActorNameException in the right place
We don't really need "remainingRetries", and I'm not sure about
"errorCount", as we've decided to retry infinitely in the case of
persistor errors. THough this isn't a persistor error, not sure there
should be a retry limit for this, either (barring some race condition?).

This can be further refined to only test for the "name still reserved"
case (this exception can mean other things, e.g. programmer error) - see
https://github.com/thatdot/quine-plus/pull/1793

GitOrigin-RevId: e41b77e61590c52bc66bee6ea1daf64aa7d77250