Unify code path whether or not a snapshot is present.
* Restore MultipleValues SQ states after cancelled node shutdown

* Apply style suggestions from feedback

* Unify code path whether or not a snapshot is present.

Restore the old behavior of having the same code (apart from the
getOrElse empty) whether or not a snapshot was found in the persistor.

Also remove the wrapping all persitor errors with
`NodeWakeUpFailedException` - the scaladoc for it states it represents a
non-recoverable error, and our `NodeAndShardSupervisorStrategy`'s
handling of it confirms that interpretation. That's not appropriate for
e.g. a transient persistor error we should retry on.
The only other current uses of that exception are for deserialization
errors - which are indeed non-recoverable.

Co-authored-by: emanb29 <ethan@thatdot.com>
GitOrigin-RevId: a8f2cd5d8f74a72ff1420f7ec51f2d0cb31b6e3c