Get rid of a couply `lazy val`s and a `getOrElse(false)`
`Option.getOrElse(false)` can generally be replaced by replacing the
prior `Option` predicate with `exists`.

lazy vals aren't free - they involve locks and caching (e.g. I think a
`def` is probably cheaper if the "cached" value is only going to be used
once). In this case where they're method local variables, apparently
they involve a `VolatileByteRef`. Was a bit confused by the need for
"volatile" in a local variable like this, but I suppose we could always
fork off a thread from here that references the `lazy val`?
https://stackoverflow.com/a/17642772/283260
Anyway, they're a bit complicated underneath (compared to e.g. a direct
`val` or even a `def`), and while normally we shouldn't have to think
about that complication, since scalac claims to take care of it for us,
I've hit scalac bugs before, in cases where the generating of the
bytecode gets tricky. And we've been bitten by the impl of lazy
initialization here before, when we tried `synchronized(this)` on
`object` fields.

The `lazy val` is made unecessary here simply by moving it into the only
scope where it's used.

GitOrigin-RevId: 76b51e4f692fe9e01755806d08ec6f9a6bd3180d