Kafka Properties merge efficenties
Added additional configurations that allows users to use Kafka that is secured with SASL (PLAIN) as an ingest stream. Currently only non secured kafka topics are able to be configured in the recipe.yaml file.

Here is an example of how you would use the new configuration options. saslAuthentication is optional and does not need to be set, so existing flow of unsecured kafka brokers should work as it did before.

```yaml
ingestStreams:
  - type: KafkaIngest
    topics:
      - upserts
    bootstrapServers: <server>
    saslAuthentication:
      type: Plain
      jaasConfig: org.apache.kafka.common.security.plain.PlainLoginModule required username="<user>" password="<pass>";
      saslMechanism: PLAIN
    securityProtocol: SASL_SSL
```

Closes thatdot/quine#23

Co-authored-by: charliemenke <charliemenke@me.com>
GitOrigin-RevId: 6424f3ce0dc5d2b8487359cda97a017547597fea