Move non-blocking computation in MapDB off the blocking dispatcher
* Move non-blocking computation in MapDB off the blocking dispatcher

And onto the default fork-join pool.

Also remove casts in favor of explicit calls to `Long` `box` / `unbox`.

The RocksDB persistor doesn't exactly have this issue, as calls there
are done by wrapping the entire method with Future on the blocking
dispatcher. So means other computation is running on a blocking thread -
but at least it's not re-dispatching to the blocking threadpool, and
instead continuing execution on the same thread (similiar to if
ExecutionContexts.parasitic and been used).

GitOrigin-RevId: 97145fbd9e77c605ab4f6282fb5bbb5629940679