Check for the "actor name is not unique!" error message for our "ActorNameStillReserved" case.
* Check for the "actor name is not unique!" error message for our "still reserved" case.

That particular exception is thrown in a variety of cases in Akka:
https://github.com/akka/akka/search?q=%22throw+InvalidActorNameException%22

The one we're interested in for our ActorNameStillReserved case is
thrown in 2 different places in this file:
https://github.com/akka/akka/blob/7abc41cf4e7e8827393b181cd06c5f8ea684e696/akka-actor/src/main/scala/akka/actor/dungeon/ChildrenContainer.scala#L134

Since Akka didn't give a way of distinguishing these cases besides
looking at the text of the message, this is apparently the best we can
do.

The other cases I imagine would be the result of static programmer
error, or perhaps a quine id that's not valid as an actor name?
These exceptions that lead to the ActorNameStillReserved case are never
seen or heard from again - I don't imagine it'd be fun to debug if other
cases were getting swallowed under there. These other cases should be
treated as "unexcpected exceptions" and have some info about them logged
(e.g. stack trace, error message).

Also removed unused exception field in ActorNameStillReserved, and
changed log level of "can't wake up actor" to error - since we currently
shut down the actor system because of that case.

* Use the case class unapply of InvalidActorNameException.

GitOrigin-RevId: 51ac8dc01bc5d58b80365f78af971aff3a4bdfb6