QU-786: Adds a jsonOrYamlRequest to endpoints4s, which can be used in place of jsonRequest
* Add a YAML => String method

Special case application/yaml content-type to add a YAML-formatted
example next to the schema. Uses the same exaple, just formatted as a
YAML string.
Ideally this is something that would go upstream, but endpoints4s
doesn't seem to support adding an example as a sibling to the schema at
the moment. As we've already forked their schema to OpenAPI json
renderer, this works for now as a place to put that.

* Add "WithExample" variants of yamlRequest

* Add YAML support to rest of JSON PUT / POST endpoints

* Add TODO about upstreaming support to endpoints4s.

It's using a java.io.InputStream, so it stands to reason there's
blocking IO somewhere - but the source here is an Akka HTTP request
body...
There's also Akka's `Dispatchers.DefaultBlockingDispatcherId, as an
alternative to `QuineDispatchers.blockingDispatcherName`?

* Make YAML last in the orElse so JSON comes first in the opeapi drop-down.

* Break up yaml request parser a little, make helper method.

* Add comments about where the blocking occurs for the InputStream

GitOrigin-RevId: b8372d18c10906b93f521bcb4b28e0329d415d56