DNS discovery: QU-695
* Default to using your own ip.

* Do a DNS lookup on seed address

* Propgate cluster joining config through to cluster endpoint

* Cleanup: remove dead code, commented code, printlns

* Still join "cluster" in singleton node case, too.

I suppose that could also be added to the Akka config, instead of
calling Cluster.joinSeedNodes?

Also cleared up some 2.13 warnings about "inferred Object".

* Handle the (default) single-node case of Quine Enterprise

Defaults to using itself as a seed node and joining that.

* Fix 2.13 "A type was inferred to by Any" warnings.

Typesafe Config doesn't really give you a typesafe DSL for constructing
these values...

* Fix test compile

* Fix documented config w/ new default values.

* Add Host and Port "AnyVal" wrappers for config special handling.

So we can have ConfigConvert instances that handle translating special
values like "<getHosAddress>" and 0 when reading from config (String and
Int would be to general of a place to put instance for that stuff).

Removed special constructors for `QuineHost` and `WebServerConfig` (this
is handled by the above config readers now. And we don't have a use-case
for that in `QuineHost` - that option is a config-only thing, and the
only place you can specify `QuineHost`s in config is in seed addresses,
where that kind of thing wouldn't make sense).

* Use passed-in URL value

Since thats what this whole block is triggered on.

* Update comments in documented config.

Explains about new dns-entry option.

* Use a condition (Some) instead of lack of exceptions to denote success.

This way retries based on the expected "No DNS entry found (yet)" are
different than errors. We probably don't want it just retrying forever
if there's an actual error.

Will retry 10,000 times before giving up. So that means the system will
wait up to like 83 minutes before giving up on a seed node(s) coming up.

GitOrigin-RevId: 39b061faf77623eee5f09d6c38a96eec9e4f99e0