Simplify entrypoints to cypher functionality
The primary aim of this PR is to simplify and reduce the number of invariants a user must maintain to "correctly" start a cypher query. To this end, the catch-all "query" function has been mostly-removed (now used only for complex continuations, like transferring ownership of a query to a SkipOptimizingActor) and has been replaced with:
- `CypherOpsGraph.cypherOps.query`, taking an `Anywhere` (i.e., OK to run off-node) query, a timestamp at which to run, and parameters
- `CypherBehavior.runQuery`, taking any query and parameters

The former is used in most cases. The latter is what is needed to support AlgorithmBehavior, and makes a cleaner path to using cypher from within the graph in general

GitOrigin-RevId: 4035b173d0a666144c23d2d4a345030f75babb77