Cypher: better detection of query anchors
Some queries that should have been fully anchored weren't. The test that
is added is an example: prior to this commit, that test fails since there
is an all-node-scan

The issue boils down to not openCypher rewriting equality `x = y` into
`x IN [y]` or `y IN [x]` and `quine-cypher` not recognizing all of those
in the same way.

Closes thatdot/quine#28

Co-authored-by: Alec Theriault <alec.theriault@gmail.com>
GitOrigin-RevId: 9cae8819c7a670cef12b6269b85b00bcc978b02a