Move test using embedded cassandra to src/it
The "it" stands for "integration test".
So now it is run with `IntegrationTest/test`. `IntegrationTest` is just
a new SBT scope with the same keys as the `Test` scope.
See https://www.scala-sbt.org/1.x/docs/Testing.html#Integration+Tests
Launching a forked Cassandra to run tests against is stretching the
bounds of what is considered a unit test.
embedded-cassandra doesn't work on JVMs, and I suspect it may be the
reason GitHub keeps killing our CI runs.
We don't need to run this test on every PR; we only need it for testing
code that interacts with Cassandra.
Also, these kinds of tests (that interact with external resources) can
be run in an environment that already has those resources spun up. E.g.
on a laptop, or AWS CodeBuild runs can be run in a specific VPC, i.e.
one that already has those spun up, like a staging environment.

GitOrigin-RevId: df87bb36484a0808f879cbbbd9c2fdd6036533b9