Use msgpack / unpacker directly
instead of allocating and intermediate byte array and wrapping it
with a ByteBuffer to do the same thing.

Less code, and more efficient.

Also use Int instead of Long for java.time.LocalDate. Int goes up to the
year 5,881,580. Long is probably well beyond the heat death of the
universe, and throws exceptions in LocalDate for larger values.

GitOrigin-RevId: 446ed4d0af4ea5817e738086e90f5be8fd2fe1db