Pass around case class instead of tuples
It can get awkward to deal witih tuples of large number of elements - so
we make case classes that name the fields.

* Return case class directly from the Cassandra table, instead of an
intermediate tuple that we later map into the case class.

* Remove then-unused Cassandra helpers for selecting 3-columns at once
as a tuple. I'm thinking in the future, if we want 3 or more things
tupled together in the response from a table, we'll probably have a case
class for it.

* Use flatMap instead of map + flatten

* Use `.traverse` on the `Option` returned in the middle of flatMapping
database calls together instead of pattern matching on it and having a
`map(Some(...` and a `case _ => Future.successful(None)`. `.traverse` is
also known as `mapM` (from before Applicative and Traversable
typeclasses existed). The type signature on `Option.traverse` here is
`Option[A] => (A => Future[B]) => Future[Option[B]]`, i.e. the same
thing as `map`, but where the function passed in does an "effect" (here,
`Future`). The `Future.traverse` in the stdlib is a special case of
this, specialized to `TraversableOnce` (which doesn't include Option?)
and Future.

GitOrigin-RevId: 5c4355af0a59bd39f2ab4a314ddaf314155242b9