Append reference.conf from our project code last when merging in assembly jar
This just moves our (obfuscated) code to be at the end rather than the
beginning of the classpath that's passed off to the assembly stage. so
that our settings appear at the bottom of reference.conf, so they can
override other libraries reference.conf settings.

The setup here is that we pass in all our libraries
(`externalDependencyClasspath`) + our obfuscated code (running proguard
on the value of `exportedProducts` from our project) to sbt-assembly as
its `fullClasspath` setting.

Also removed our `reverseAppend` from Obfuscation in favor of using its
built-in `append`. Maybe that was a relic from when we ran proguard on
everything - but now that we're running proguard only on our own code,
it doesn't seem necessary to worry about the order there, as hopefully
we're not trying to override the same setting to multiple different
values between our projects in the same stack?

* Use whether its a project or library to determine what goes last

As opposed to just reversing the input list

* Rename reverseConcat to projectsLast

To reflect the new thing this strategy does

* Why would be have multiple application.confs?

AFAIK novelty is the only project w/ an appliaction.conf.
I thought we had moved away from using application.conf in favor of of
using reference.conf in our apps?

Also, why does this application.conf include both akka-overrides and
plus-akka-overrides, while quine-enterprise just includes the latter and
uses the import form quine-core's reference.conf for the former?

GitOrigin-RevId: 8a1ad7d0c15fd3ffcba6ffdb47df5605ea629627