QU-1139: purgeNode() method in Cypher to remove a node from history
* ScalaTest w/ delete at beginning

* Factor out common string between add-opens JVM args

* Added deleteSnapshots to persistence agent.

* Move persistor test using deletion into memory and cassandra-specific tests

Only places that delete are implemented so far.

Also add helpers for running a Cassandra query as a Scala Future, and
for queries that return zero or one rows (returns Option).

That can also be done using Akka Streams Sink.headOption, but adding
akka streams and an ActorSystem just for that seems like a bit of
overhead.

* Make `CassandraStatementSettings` for passing in read / write settings as a pair.
Just pairs the consistency level and timeout. Also a helper function for
applying those settings to a Statement.

* Also widen the consistency level in config from
DefaultConsistencyLevel to ConsistencyLevel. The former is an enum that
is the sole implemententation of the interface ConsistencyLevel. The
interface only exists for extensibility reasons. But there's no reason
we have to be restrictive about it - we don't care - we're just a
pass-through for this setting.

* Update parameters on CassandraPersistorPlus builder

* Re-skip a test that fails on CI w/ InMemoryPersistor

* Wrap persistor exceptions from new delete endpoints

* Added a purgeNode test from Cypher

Also added other delete methods to InMemoryPersistor, which is what the
Cypher tests use.

* Update public/quine-cypher/src/main/scala/com/thatdot/quine/compiler/cypher/Procedures.scala

Co-authored-by: Ethan Bell <ethan@thatdot.com>

---------

Co-authored-by: Ethan Bell <ethan@thatdot.com>
GitOrigin-RevId: ae236c180c6a0a298debc846db2bcbd4c17d3feb