Update docs about metrics and liveness / readiness under admin API
* Update docs about metrics and liveness / readiness under admin API

* Clarify the time measurement of timers.

* Clarify the difference between liveness vs readiness, and remove talk
about containers / kubernetes (these "health check" enpoints are general
and appropriate for use from any normal process manager / load balancer
/ anything routing requests to us).

* Remove unused boolean / 503 response from liveness endpoint

There is currently no possible way for this endpoint to return a 503, so
telling the user that is a possible option seems disingenuous. It can of
course return a 500 response (on uncaught exception) like all endpoints.
Since this endpoint is a no-op, that scenarios seems unlikely (unless
you have a misconfigured system).

The boolean parameter offers the possibility we could take some process
decision internally in the future on that.
However, as the point of this endpoint is for detecting if the process
is hung / non-responsive and needs to be SIGTERM / SIGKILL'd, having a
decision made in our applicaiton logic on that doesn't seem to make a
lot of sense. That is, our applicaiton logic can't really answer that
question for itself: "Is the process I'm running as hung and needs to be
killed?". So a simple no-op endpoint to poll (as currently implemented)
I think is the most sensible thing for this.

GitOrigin-RevId: 562f26d9d7ab84597aad2055a161088c9a0ec7d4