Store offsets in datetimes
* Store offsets in datetimes

Instead of normalizing every timestamp to UTC.

This means using java.time.OffsetDateTime for QuineValue's DateTime
representation, instead of java.time.Instant.

An extra byte is used to store the offset, which is represented in
15-minute increments between -12:00 and +14:00

The storage space used changes from either 32, 64, or 96 bits (depending
on the range & precision of the timestamp) to 72 or

The maximum date range has been reduced from +/- 1,000,000,000 years
(the range for java.time.Instant - dunno if ZonedDateTime imposes its
own limits) to about +/- 5,881,580 years
`java.time.LocalDate.ofEpochDay(Int.MaxValue)` in the interest of not
requiring a 64 bits instead of 32 bits to store the date.

The ability to read the old format (Instants, via msgpacks's built-in
format for them) is left intact. New data will be written in the new
format.

* Add a Pickler for OffsetDateTime

We're using this now for our QuineValue.DateTime representation.

* Update protobuf and reify time uses of datetime.

GitOrigin-RevId: 53bc2c9e8b3ea8dc31c7a69e5a269cc6e51f6878