Make Hashable derivation not infinitely recurse on missing instance.
* Make Hashable derivation not infinitely recurse on missing instance.

Having the compiler indefinitely hang makes it hard to see where the
issue is - if the compiler's not giving you any output.

Also added a `Hashable` instance for `OffsetTime`.

Removed the variance annotation from `Hashable`, and removed the single
implicit parameterized over all subtypes of `Iterable`, in favor of
implicits for specific instances (implemented by that general iterable
impl), in keeping with the patterns seen in generic derivation in other
libraries. In my experience, subtyping and implicit search don't mix
well, and it's probably best to keep implicit search as specific and
constrained as possible. And it's suspicious that the prior iterable1
impl seemed to require a `Lazy` on its element `Hashable` parameter.

Also removed `iterable2` (Maps can be hashed by typing them as an
Iterable of tuples and using the existing iterable hash implementation).

Rewrote the shapeless Generic derivation for Hashable using its
`TypeClass` helper class. The benefits for this helper aren't as much if
you're just using Generic instead of LabelledGeneric - it was mainly to
make sure we were following a "known good" pattern, as the previous impl
seemed to have an issue w/ infinitely recursing at compile time if an
instance was missing.

* Switch UnitSq back to being a case class (rather than an object)

For cursed "concrete val in superclass" / shapeless reasons.

GitOrigin-RevId: 2cb020dd5ebf4e728200b45c802e09b46300a77e